+++ src/main/nextflow/co2footprint/Recorders/SessionTraceRecorder.groovy
@@ -1,11 +1,15 @@
 package nextflow.co2footprint.Recorders
 
+import groovy.util.logging.Slf4j
 import nextflow.Session
 import nextflow.trace.TraceRecord
 import oshi.SystemInfo
+import oshi.driver.linux.proc.ProcessStat
 import oshi.hardware.CentralProcessor
 import oshi.software.os.OSProcess
 import oshi.software.os.OperatingSystem
+import oshi.software.os.linux.LinuxOperatingSystem
+import oshi.util.tuples.Triplet
 
 import java.lang.management.ManagementFactory
 import java.lang.management.RuntimeMXBean
@@ -16,6 +20,7 @@ import com.sun.management.OperatingSystemMXBean
  * A Recorder of trace values for a Nextflow session, which can be attached after startup
  * to capture timepoints before workflow invocation.
  */
+@Slf4j
 class SessionTraceRecorder {
     // OSHI info handles
     private final RuntimeMXBean          runtimeBean = ManagementFactory.getRuntimeMXBean()
@@ -30,7 +35,7 @@ class SessionTraceRecorder {
     // Process information
     private int pid
     private OSProcess process
-    private Map<Integer, OSProcess> allProcesses = [(pid): process].asSynchronized()
+    private Map<Integer, OSProcess> allProcesses = [:].asSynchronized()
 
     // Aggregation
     final List<MemorySample> samples = [].asSynchronized() as List<MemorySample>
@@ -42,6 +47,7 @@ class SessionTraceRecorder {
     void start() {
         pid = runtimeBean.pid as int
         process = os.getProcess(pid)
+        allProcesses.put(pid, process)
 
         sessionRecord.putAll(
                 [
@@ -89,6 +95,27 @@ class SessionTraceRecorder {
         List<OSProcess> processList = allProcesses.values() as List<OSProcess>
         processList.each({ OSProcess p -> p.updateAttributes() })
 
+        Map<ProcessStat.PidStat, Long> rootStats = getPidStats(pid)
+
+        // Determine CPU usage
+        Double cpuUsage
+        if (rootStats != null) {
+            // Accumulate the running ticks of root including waiting time for children
+            Long cpuTime = rootStats.get(ProcessStat.PidStat.UTIME) + rootStats.get(ProcessStat.PidStat.STIME) +
+                    rootStats.get(ProcessStat.PidStat.CUTIME) + rootStats.get(ProcessStat.PidStat.CSTIME)
+            // Convert from jiffies to ms
+            cpuTime = (cpuTime * 1000 / LinuxOperatingSystem.getHz()) as Long
+
+            // Calculate elapsed time since process start
+            Long elapsedTime = endTimestamp - process.startTime
+
+            cpuUsage = cpuTime / elapsedTime
+            log.debug("Calculated CPU usage for PID ${pid}: ${cpuUsage}")
+        }
+        else {
+            cpuUsage = processList.collect({ OSProcess p -> p.getProcessCpuLoadCumulative() }).sum() as double
+        }
+
         sessionRecord.putAll(
                 [
                         status:         'COMPLETED',
@@ -96,21 +123,21 @@ class SessionTraceRecorder {
                         duration:       endTimestamp - (sessionRecord.get('submit') as long),
                         realtime:       runtimeBean.uptime,
                         memory:         Runtime.getRuntime().maxMemory(),
-                        '%cpu':         (processList.sum({ OSProcess p -> p.getProcessCpuLoadCumulative() }) as double) * 100,
-                        read_bytes:     processList.sum({ OSProcess p -> p.bytesRead}) as Long,
-                        write_bytes:    processList.sum({ OSProcess p -> p.bytesWritten}) as Long,
-                        vol_ctxt:       processList.sum({ OSProcess p -> p.minorFaults}) as Long,
-                        inv_ctxt:       processList.sum({ OSProcess p -> p.majorFaults}) as Long,
+                        '%cpu':         cpuUsage * 100,
+                        read_bytes:     processList.collect({ OSProcess p -> p.bytesRead}).sum() as Long,
+                        write_bytes:    processList.collect({ OSProcess p -> p.bytesWritten}).sum() as Long,
+                        vol_ctxt:       processList.collect({ OSProcess p -> p.minorFaults}).sum() as Long,
+                        inv_ctxt:       processList.collect({ OSProcess p -> p.majorFaults}).sum() as Long,
                 ] 
         )
 
         if (samples) {
             sessionRecord.putAll(
                     [
-                            rss:            samples.average({ MemorySample sample -> sample.rssBytes}) as Long,
-                            vmem:           samples.average({ MemorySample sample -> sample.virtualMemoryBytes}) as Long,
-                            peak_rss:       samples.max({ MemorySample sample -> sample.rssBytes}),
-                            peak_vmem:      samples.max({ MemorySample sample -> sample.virtualMemoryBytes}),
+                            rss:            samples.collect({ MemorySample sample -> sample.rssBytes}).average() as Long,
+                            vmem:           samples.collect({ MemorySample sample -> sample.virtualMemoryBytes}).average() as Long,
+                            peak_rss:       samples.collect({ MemorySample sample -> sample.rssBytes}).max() as Long,
+                            peak_vmem:      samples.collect({ MemorySample sample -> sample.virtualMemoryBytes}).max() as Long,
                     ]
             )
         }
@@ -124,6 +151,27 @@ class SessionTraceRecorder {
     void stop() {
         timer.cancel()
         timer.purge()
+    }
+
+    /**
+     * Attempts to fetch the PID stats for the current process.
+     *
+     * @param pid The PID for which to fetch the stats
+     * @return Map of PID stats, or null if the information could not be retrieved (e.g. due to permissions or OS)
+     */
+    Map<ProcessStat.PidStat, Long> getPidStats(Integer pid) {
+        LinuxOperatingSystem
+        if (os instanceof LinuxOperatingSystem) {
+            try {
+                Triplet<String, Character, Map<ProcessStat.PidStat, Long>> stats = ProcessStat.getPidStats(pid)
+                return stats.getC()
+            }
+            catch (Exception e) {
+                log.trace("Failed to get process stats for PID ${pid}: ${e.message}")
+            }
+        }
+
+        return null
     }
 
     /**
@@ -145,8 +193,8 @@ class SessionTraceRecorder {
                     timestamp: System.currentTimeMillis(),
 
                     // Memory
-                    rssBytes: activeProcesses.sum({ OSProcess p -> p.residentSetSize}) as Long,
-                    virtualMemoryBytes: activeProcesses.sum({ OSProcess p -> p.virtualSize}) as Long,
+                    rssBytes: activeProcesses.collect({ OSProcess p -> p.residentSetSize}).sum() as Long,
+                    virtualMemoryBytes: activeProcesses.collect({ OSProcess p -> p.virtualSize}).sum() as Long,
             )
 
             samples.add(sample)